rewritted Legend component for ts files
